+++ RELEASENOTES.md
@@ -2,6 +2,18 @@
 # React Components: design-system-react
 # Release notes
 
+## Release 0.3.4
+
+**MAJOR CHANGES - SLDSPageHeader**
+- Make field truncation optional.
+- If field truncates, then show PopoverTooltip on hover/focus to view full field text.
+
+**Bug Fix**
+- Lookup component had a bug where updating the selectedItem prop did not work. Now it does. To clear the item, pass in -1.
+- Focus on search input when App Launcher opens
+- No longer renders a close button if modal type is prompt.
+- Lookup search icon default should be on right side.
+
 ## Release 0.3.2
 
 **MAJOR CHANGES**